Add -c Option, Start Web-Form
Added the -c option, which will allow for a SpamAssassin meta-rule to be generated which will allow exemptions to the custom rule being created, even for systems not supporting look-aheads. Starting a web-form (GUI) for this tool to run in. It will be a modular ReactJS app that can be dropped onto any web-server in any location. Also added a few more first-names from a list of common first names. Fixing smaller typos/notes, altering program output based on initial options. Updated README.