1.0 - Initial Release of Web Form
Made the design mobile-device compatible so it can now be used from different non-standard viewports as well. The form will flex from 300px to 610px or so, depending on the viewport width. Built the project and updated the README.md doc for the web-form version. The 'build' directory is added to the project so that prospective users of the tool can have an easy-to-deploy drag-n-drop shipment of the script in a web form native to almost any browser. Lots of minor QOL improvements and bug-fixes as well.