Centralizing CSS & New Skin
Centralized the variable definitions for the CSS files so that colors can easily be defined for the tool. Removed Barracuda branding (as I don't want this project affiliated with them entirely) and set the default skin to a smooth "peach" color. Also trying to remove other Barracuda associations from the project as I want it to be applicable EVERYWHERE. Also added a disclaimer for legal stuffs, just in case. You never know.