New Form Options
In tradition nub style, I'm adding this as a bulk update (less for VC, more for archiving/synchronization). I should have tended to this garden much sooner.

Added the ability to dynamically add multple titles/honorifics to a name. Also gave the form an ability to control the threshold for typosquatting protection (as well as making that toggleable). Added the option for slicing the last name, thereby making the entirety of the last name except the initial an optional piece. Rearranged some checkboxes and added dynamic form injection so that some pieces are only visible if the option is selected.

Improved the operation of the RegexGenerator class to handle these new parameters, and also offloaded the first-name items to a separate JSON resource file.

Lots of very good changes, more to come whenever.